requirements.txt
matplotlib
pyqt6
numpy
numba
opencv-python
attractor-tools